Enable logging by default.
Since all Mahtanar boards use the ESP32-C3, logging is by default
connected to the USB_SERIAL_JTAG port, not either of the hardware
UARTs. This means there is no conflict between logging and using
both hardware UARTs to connect to the heatpump and thermostat.

The configs for the boards can be updated to remove the section which
disables the logger, and then the README no longer needs to include
instructions for optionally enabling it.